+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [Unreleased]
+
+### Fixed
+
+- `Innmind\Filesystem\File\Content::ofChunks()->size()` no longer load more than one chunk size in memory
+
 ## 7.5.0 - 2024-03-16
 
 ### Changed